Chg:Shifted the milestone functionality to the checkpoint module
