+++ checkpoint/types/msg.go
@@ -313,3 +313,90 @@ func (msg MsgCheckpointNoAck) ValidateBasic() sdk.Error {
 
 	return nil
 }
+
+///#######Milestone Module############
+
+var _ sdk.Msg = &MsgMilestone{}
+
+// MsgMilestone represents milestone
+type MsgMilestone struct {
+	Proposer    types.HeimdallAddress `json:"proposer"`
+	StartBlock  uint64                `json:"start_block"`
+	EndBlock    uint64                `json:"end_block"`
+	RootHash    types.HeimdallHash    `json:"root_hash"`
+	BorChainID  string                `json:"bor_chain_id"`
+	MilestoneID string                `json:"milestone_id"`
+}
+
+// NewMsgMilestoneBlock creates new milestone message using mentioned arguments
+func NewMsgMilestoneBlock(
+	proposer types.HeimdallAddress,
+	startBlock uint64,
+	endBlock uint64,
+	roothash types.HeimdallHash,
+	borChainID string,
+	milestoneID string,
+) MsgMilestone {
+	return MsgMilestone{
+		Proposer:    proposer,
+		StartBlock:  startBlock,
+		EndBlock:    endBlock,
+		RootHash:    roothash,
+		BorChainID:  borChainID,
+		MilestoneID: milestoneID,
+	}
+}
+
+// Type returns message type
+func (msg MsgMilestone) Type() string {
+	return "milestone"
+}
+
+func (msg MsgMilestone) Route() string {
+	return RouterKey
+}
+
+// GetSigners returns address of the signer
+func (msg MsgMilestone) GetSigners() []sdk.AccAddress {
+	return []sdk.AccAddress{types.HeimdallAddressToAccAddress(msg.Proposer)}
+}
+
+func (msg MsgMilestone) GetSignBytes() []byte {
+	b, err := ModuleCdc.MarshalJSON(msg)
+	if err != nil {
+		panic(err)
+	}
+
+	return sdk.MustSortJSON(b)
+}
+
+func (msg MsgMilestone) ValidateBasic() sdk.Error {
+	if bytes.Equal(msg.RootHash.Bytes(), helper.ZeroHash.Bytes()) {
+		return hmCommon.ErrInvalidMsg(hmCommon.DefaultCodespace, "Invalid rootHash %v", msg.RootHash.String())
+	}
+
+	if msg.Proposer.Empty() {
+		return hmCommon.ErrInvalidMsg(hmCommon.DefaultCodespace, "Invalid proposer %v", msg.Proposer.String())
+	}
+
+	if msg.StartBlock >= msg.EndBlock || msg.EndBlock == 0 {
+		return hmCommon.ErrInvalidMsg(hmCommon.DefaultCodespace, "Invalid startBlock %v or/and endBlock %v", msg.StartBlock, msg.EndBlock)
+	}
+
+	return nil
+}
+
+// GetSideSignBytes returns side sign bytes
+func (msg MsgMilestone) GetSideSignBytes() []byte {
+	// keccak256(abi.encoded(proposer, startBlock, endBlock, rootHash, accountRootHash, bor chain id))
+	borChainID, _ := strconv.ParseUint(msg.BorChainID, 10, 64)
+
+	return appendBytes32(
+		msg.Proposer.Bytes(),
+		new(big.Int).SetUint64(msg.StartBlock).Bytes(),
+		new(big.Int).SetUint64(msg.EndBlock).Bytes(),
+		msg.RootHash.Bytes(),
+		new(big.Int).SetUint64(borChainID).Bytes(),
+		[]byte(msg.MilestoneID),
+	)
+}
