--- milestone/client/cli/flags.go
@@ -1,14 +1,0 @@
-package cli
-
-const (
-	FlagProposerAddress = "proposer"
-	FlagFromAddress     = "from"
-	FlagStartBlock      = "start-block"
-	FlagEndBlock        = "end-block"
-	FlagRootHash        = "root-hash"
-	FlagBorChainID      = "bor-chain-id"
-	FlagMilestoneNumber = "number"
-	FlagHeaderNumber    = "header"
-	FlagAutoConfigure   = "auto-configure"
-	FlagMilestoneID     = "milestone-id"
-)
