--- milestone/client/cli/tx.go
@@ -1,118 +1,0 @@
-package cli
-
-import (
-	"fmt"
-	"strconv"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/context"
-	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-
-	hmClient "github.com/maticnetwork/heimdall/client"
-	"github.com/maticnetwork/heimdall/helper"
-	types "github.com/maticnetwork/heimdall/milestone/types"
-	hmTypes "github.com/maticnetwork/heimdall/types"
-)
-
-var logger = helper.Logger.With("module", "milestone/client/cli")
-
-// GetTxCmd returns the transaction commands for this module
-func GetTxCmd(cdc *codec.Codec) *cobra.Command {
-	txCmd := &cobra.Command{
-		Use:                        types.ModuleName,
-		Short:                      "milestone transaction subcommands",
-		DisableFlagParsing:         true,
-		SuggestionsMinimumDistance: 2,
-		RunE:                       hmClient.ValidateCmd,
-	}
-
-	txCmd.AddCommand(
-		client.PostCommands(
-			SendMilestoneTx(cdc),
-		)...,
-	)
-
-	return txCmd
-}
-
-// SendMilestoneTx send milestone transaction
-func SendMilestoneTx(cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "send-milestone",
-		Short: "send milestone to tendermint",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-
-			// bor chain id
-			borChainID := viper.GetString(FlagBorChainID)
-			if borChainID == "" {
-				return fmt.Errorf("bor chain id cannot be empty")
-			}
-
-			// get proposer
-			proposer := hmTypes.HexToHeimdallAddress(viper.GetString(FlagProposerAddress))
-			if proposer.Empty() {
-				proposer = helper.GetFromAddress(cliCtx)
-			}
-
-			//	start block
-			startBlockStr := viper.GetString(FlagStartBlock)
-			if startBlockStr == "" {
-				return fmt.Errorf("start block cannot be empty")
-			}
-			startBlock, err := strconv.ParseUint(startBlockStr, 10, 64)
-			if err != nil {
-				return err
-			}
-
-			//	end block
-			endBlockStr := viper.GetString(FlagEndBlock)
-			if endBlockStr == "" {
-				return fmt.Errorf("end block cannot be empty")
-			}
-			endBlock, err := strconv.ParseUint(endBlockStr, 10, 64)
-			if err != nil {
-				return err
-			}
-
-			// root hash
-			rootHashStr := viper.GetString(FlagRootHash)
-			if rootHashStr == "" {
-				return fmt.Errorf("root hash cannot be empty")
-			}
-
-			milestoneID := viper.GetString(FlagMilestoneID)
-
-			msg := types.NewMsgMilestoneBlock(
-				proposer,
-				startBlock,
-				endBlock,
-				hmTypes.HexToHeimdallHash(rootHashStr),
-				borChainID,
-				milestoneID,
-			)
-
-			return helper.BroadcastMsgsWithCLI(cliCtx, []sdk.Msg{msg})
-		},
-	}
-	cmd.Flags().StringP(FlagProposerAddress, "p", "", "--proposer=<proposer-address>")
-	cmd.Flags().String(FlagStartBlock, "", "--start-block=<start-block-number>")
-	cmd.Flags().String(FlagEndBlock, "", "--end-block=<end-block-number>")
-	cmd.Flags().StringP(FlagRootHash, "r", "", "--root-hash=<root-hash>")
-	cmd.Flags().String(FlagBorChainID, "", "--bor-chain-id=<bor-chain-id>")
-	cmd.Flags().Bool(FlagAutoConfigure, false, "--auto-configure=true/false")
-	cmd.Flags().String(FlagMilestoneID, "0000", "--milestone-id=<milestone-id>")
-
-	if err := cmd.MarkFlagRequired(FlagRootHash); err != nil {
-		logger.Error("SendCheckpointTx | MarkFlagRequired | FlagRootHash", "Error", err)
-	}
-
-	if err := cmd.MarkFlagRequired(FlagBorChainID); err != nil {
-		logger.Error("SendCheckpointTx | MarkFlagRequired | FlagBorChainID", "Error", err)
-	}
-
-	return cmd
-}
