--- milestone/querier.go
@@ -1,127 +1,0 @@
-package milestone
-
-import (
-	"encoding/json"
-	"fmt"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/maticnetwork/heimdall/common"
-	"github.com/maticnetwork/heimdall/helper"
-	"github.com/maticnetwork/heimdall/milestone/types"
-	"github.com/maticnetwork/heimdall/staking"
-	abci "github.com/tendermint/tendermint/abci/types"
-)
-
-// NewQuerier creates a querier for auth REST endpoints
-func NewQuerier(keeper Keeper, stakingKeeper staking.Keeper, contractCaller helper.IContractCaller) sdk.Querier {
-	return func(ctx sdk.Context, path []string, req abci.RequestQuery) ([]byte, sdk.Error) {
-		switch path[0] {
-		case types.QueryParams:
-			return handleQueryParams(ctx, req, keeper)
-
-		case types.QueryLatestMilestone:
-			return handleQueryLatestMilestone(ctx, req, keeper)
-
-		case types.QueryMilestoneByNumber:
-			return handleQueryMilestoneByNumber(ctx, req, keeper)
-
-		case types.QueryCount:
-			return handleQueryCount(ctx, req, keeper)
-
-		case types.QueryLatestNoAckMilestone:
-			return handleQueryLatestNoAckMilestone(ctx, req, keeper)
-
-		case types.QueryNoAckMilestoneByID:
-			return handleQueryNoAckMilestoneByID(ctx, req, keeper)
-
-		default:
-			return nil, sdk.ErrUnknownRequest("unknown auth query endpoint")
-		}
-	}
-}
-
-func handleQueryParams(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
-	bz, err := json.Marshal(helper.MilestoneLength)
-
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
-	}
-
-	return bz, nil
-}
-
-func handleQueryLatestMilestone(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
-	res, err := keeper.GetLastMilestone(ctx)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not fetch milestone", err.Error()))
-	}
-
-	if res == nil {
-		return nil, common.ErrNoMilestoneFound(keeper.Codespace())
-	}
-
-	bz, err := json.Marshal(res)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
-	}
-
-	return bz, nil
-}
-
-func handleQueryMilestoneByNumber(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
-	var params types.QueryMilestoneParams
-	if err := keeper.cdc.UnmarshalJSON(req.Data, &params); err != nil {
-		return nil, sdk.ErrInternal(fmt.Sprintf("failed to parse params: %s", err))
-	}
-
-	res, err := keeper.GetMilestoneByNumber(ctx, params.Number)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not fetch milestone", err.Error()))
-	}
-
-	if res == nil {
-		return nil, common.ErrNoMilestoneFound(keeper.Codespace())
-	}
-
-	bz, err := json.Marshal(res)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
-	}
-
-	return bz, nil
-}
-
-func handleQueryCount(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
-	bz, err := json.Marshal(keeper.GetCount(ctx))
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
-	}
-
-	return bz, nil
-}
-
-func handleQueryLatestNoAckMilestone(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
-	res := keeper.GetLastNoAckMilestone(ctx)
-
-	bz, err := json.Marshal(res)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
-	}
-
-	return bz, nil
-}
-
-func handleQueryNoAckMilestoneByID(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
-	var ID types.QueryMilestoneID
-	if err := keeper.cdc.UnmarshalJSON(req.Data, &ID); err != nil {
-		return nil, sdk.ErrInternal(fmt.Sprintf("failed to parse milestoneID: %s", err))
-	}
-	res := keeper.GetNoAckMilestone(ctx, ID.MilestoneID)
-
-	bz, err := json.Marshal(res)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
-	}
-
-	return bz, nil
-}
