--- milestone/simulation/genesis.go
@@ -1,37 +1,0 @@
-package simulation
-
-import (
-	"time"
-
-	"github.com/maticnetwork/heimdall/milestone/types"
-	hmTypes "github.com/maticnetwork/heimdall/types"
-	"github.com/maticnetwork/heimdall/types/module"
-)
-
-// RandomizedGenState return dummy genesis
-func RandomizedGenState(simState *module.SimulationState) {
-
-	startBlock := uint64(0)
-	endBlock := uint64(256)
-	rootHash := hmTypes.HexToHeimdallHash("123")
-
-	proposerAddress := hmTypes.HexToHeimdallAddress("123")
-	timestamp := uint64(time.Now().Unix())
-	borChainID := "1234"
-	milestoneID := "00000"
-
-	milestone := hmTypes.CreateMilestone(
-		startBlock,
-		endBlock,
-		rootHash,
-		proposerAddress,
-		borChainID,
-		milestoneID,
-		timestamp,
-	)
-
-	genesisState := types.NewGenesisState(
-		&milestone,
-	)
-	simState.GenState[types.ModuleName] = simState.Cdc.MustMarshalJSON(genesisState)
-}
