--- milestone/types/codec.go
@@ -1,19 +1,0 @@
-package types
-
-import (
-	"github.com/cosmos/cosmos-sdk/codec"
-)
-
-func RegisterCodec(cdc *codec.Codec) {
-	cdc.RegisterConcrete(MsgMilestone{}, "milestone/MsgMilestone", nil)
-}
-
-// ModuleCdc generic sealed codec to be used throughout module
-var ModuleCdc *codec.Codec
-
-func init() {
-	cdc := codec.New()
-	codec.RegisterCrypto(cdc)
-	RegisterCodec(cdc)
-	ModuleCdc = cdc.Seal()
-}
