--- milestone/types/events.go
@@ -1,16 +1,0 @@
-package types
-
-// Milestone tags
-var (
-	EventTypeMilestone = "milestone"
-
-	AttributeKeyProposer   = "proposer"
-	AttributeKeyStartBlock = "start-block"
-	AttributeKeyEndBlock   = "end-block"
-
-	AttributeKeyNewProposer = "new-proposer"
-	AttributeKeyRootHash    = "root-hash"
-	AttributeKeyMilestoneID = "milestone-id"
-
-	AttributeValueCategory = ModuleName
-)
