--- milestone/types/keys.go
@@ -1,18 +1,0 @@
-package types
-
-const (
-	// ModuleName is the name of the module
-	ModuleName = "milestone"
-
-	// StoreKey is the store key string for bor
-	StoreKey = ModuleName
-
-	// RouterKey is the message route for bor
-	RouterKey = ModuleName
-
-	// QuerierRoute is the querier route for bor
-	QuerierRoute = ModuleName
-
-	// DefaultParamspace default name for parameter store
-	DefaultParamspace = ModuleName
-)
